Added confirmation button for deleting post

+++ src/pages/YourPosts.tsx
@@ -13,6 +13,9 @@ import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
 import '../styles.css';
 
 export function YourPosts() {
+
+    const Swal = require('sweetalert2');
+
     const [items, setItems] = useState([]);
     useEffect(() => {
         fetch("https://localhost:7018/api/articles")
@@ -25,11 +28,29 @@ export function YourPosts() {
             )
     }, [])
 
- const deletePost = (id: string) => {
-        fetch("https://localhost:7018/api/articles/" + id, { method: 'DELETE' })
-        setItems((items) => items.filter((article: any) => article.id !== id))
+ const deletePost = (id: string) =>{
+
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result: any) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Deleted!',
+            'Article has been deleted.',
+            'success'
+          )
+          fetch("https://localhost:7018/api/articles/" + id, { method: 'DELETE' })
+          setItems((items) => items.filter((article: any) => article.id !== id))
+        }
+      })
     }
-    
+
     return (
         <Container className="paperStack" maxWidth="md" >
             <h2 className="hover1" style={{ fontFamily: "Cabin Sketch", fontSize: 50, height: 44, width: 300 }}> <DoneOutlineIcon sx={{ color: "black" }} />Your Posts
